requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
asyncpg==0.30.0
attrs==25.3.0
boto3==1.40.31
botocore==1.40.31
certifi==2025.8.3
cssselect==1.3.0
frozenlist==1.7.0
git-filter-repo==2.47.0
greenlet==3.2.4
idna==3.10
jmespath==1.0.1
lxml==6.0.1
multidict==6.6.4
packaging==25.0
parsel==1.10.0
playwright==1.55.0
playwright-stealth==2.0.0
propcache==0.3.2
pyee==13.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
s3transfer==0.14.0
six==1.17.0
SQLAlchemy==2.0.43
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
w3lib==2.3.1
yarl==1.20.1
